Add files via upload
Here are some things I've created. If you look at the Design view, this will show what the website would show you. The code can be seen in the solution explorer, go to the default.aspx, then expand it and the two files there are the entire project. One is the web view, the other is the C# code. The split view of the web page will show the HTML and CSS done.